# llss_openapi.yaml
openapi: 3.1.0
info:
  title: Low Level Screen Service (LLSS) API
  version: 0.2.0
  description: >
    The Low Level Screen Service (LLSS) brokers communication between
    e-Ink display devices and High Level Screen Service (HLSS) instances.
    It manages authentication, frame storage, diffing, and device orchestration.
    
    
    ## Authentication
    
    LLSS uses JWT-based authentication with two token types:
    
    - **Access Token** (short-lived, 1 day): Used for all authenticated API calls
    - **Refresh Token** (long-lived, 30 days): Stored securely on device, used to obtain new access tokens
    
    
    ### Device Authentication Flow
    
    1. **Registration**: Device calls `POST /auth/devices/register` with hardware_id and display capabilities.
       Returns `device_id` and `device_secret`. Device stores these in non-volatile memory (EEPROM/SPIFFS).
       Device status is set to `pending`.
    
    2. **Admin Authorization**: Administrator reviews pending devices via admin interface and authorizes them.
       Device status changes to `authorized`.
    
    3. **Get Refresh Token**: Device calls `POST /auth/devices/token` with `hardware_id` and `device_secret`.
       If authorized, returns a 30-day refresh token.
    
    4. **Get Access Token**: Device calls `POST /auth/devices/refresh` with refresh token in Authorization header.
       Returns a 1-day access token.
    
...
    
...
       If refresh fails, device re-authenticates from step 3.
    
...
servers:
...
security:
  - BearerAuth: []
...
paths:
...
  /auth/devices/register:
    post:
      tags: [Authentication]
      summary: Register a new device
      description: >
        Registers a new physical device with LLSS. This is an **unauthenticated** endpoint.
        
        
        The device provides its hardware identifier and display capabilities.
        LLSS creates a new device record with `pending` status and returns:
        - `device_id`: Unique identifier for this device
        - `device_secret`: Secret key for authentication (store securely!)
        
        
        The device must wait for admin authorization before it can obtain tokens.
        Store `device_secret` in non-volatile memory (EEPROM/SPIFFS) - it cannot be retrieved later.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceRegistrationRequest"
      responses:
        "201":
          description: Device registered successfully (pending authorization)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceRegistrationResponse"
        "409":
          description: Device with this hardware_id already exists
...
  /devices/register:
    post:
      tags: [Devices]
      summary: Register a new physical device (DEPRECATED)
      deprecated: true
      description: >
        **DEPRECATED**: Use `POST /auth/devices/register` instead.
        
        
        This endpoint is kept for backwards compatibility but will be removed
        in a future version. The new auth system uses JWT tokens.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceRegistration"
      responses:
        "201":
          description: Device registered (pending authorization)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceRegistrationResponseLegacy"
  /devices/{device_id}/state:
    get:
      tags: [Devices]
      summary: Device heartbeat and action polling
      description: >
        Core polling endpoint used by devices to report status.
        LLSS responds with the next action to perform.
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
        - name: last_frame_id
          in: query
          required: false
          schema:
            type: string
        - name: last_event_id
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Device action response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceStateResponse"
  /devices/{device_id}/frames/{frame_id}:
    get:
      tags: [Devices, Frames]
      summary: Fetch rendered frame data
      description: >
        Returns raw framebuffer data ready to be written to the e-Ink display.
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
        - name: frame_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Raw framebuffer payload
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
  /devices/{device_id}/inputs:
    post:
      tags: [Devices, Inputs]
      summary: Submit input events from a device
      description: >
        Sends button presses or other input events from the device to LLSS.
        LLSS routes them to the active HLSS instance.
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InputEvent"
      responses:
        "202":
          description: Input event accepted
  /instances:
    post:
      tags: [Instances]
      summary: Create a new HLSS instance
      description: >
        Creates a new logical HLSS instance (e.g. a chess game or HA dashboard).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstanceCreate"
      responses:
        "201":
          description: Instance created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Instance"
  /instances/{instance_id}/frames:
    post:
      tags: [Instances, Frames]
      summary: Submit a new logical frame
      description: >
        HLSS submits a newly rendered frame (PNG).
        LLSS stores, diffs, and schedules device refreshes.
      parameters:
        - name: instance_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          image/png:
            schema:
              type: string
              format: binary
      responses:
        "201":
          description: Frame accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FrameCreateResponse"
  /instances/{instance_id}/notify:
    post:
      tags: [Instances]
      summary: Notify LLSS of instance state change
      description: >
        Notifies LLSS that the instance state has changed and a new frame
        may be available or should be requested.
      parameters:
        - name: instance_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "202":
          description: Notification accepted
  /instances/{instance_id}/inputs:
    post:
      tags: [Instances, Inputs]
      summary: Receive forwarded input events
      description: >
        LLSS forwards device input events to the active HLSS instance.
      parameters:
        - name: instance_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InputEvent"
      responses:
        "200":
          description: Input processed

  # ============================================================
  # Admin Endpoints - Device Authorization
  # ============================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token obtained from `/auth/devices/refresh`.
        Valid for 1 day. Include in Authorization header as `Bearer {token}`.
    
    RefreshToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT refresh token obtained from `/auth/devices/token`.
        Valid for 30 days. Used only for `/auth/devices/refresh` endpoint.
  schemas:
...
    DeviceRegistrationRequest:
      type: object
      description: Request to register a new device
      required: [hardware_id, firmware_version, display]
      properties:
        hardware_id:
          type: string
          description: Unique hardware identifier (e.g., ESP32 chip ID)
          example: "ESP32_A1B2C3D4"
        firmware_version:
          type: string
          description: Current firmware version
          example: "1.0.0"
        display:
          $ref: "#/components/schemas/DisplayCapabilities"
    DeviceRegistrationResponse:
      type: object
      description: Response after successful device registration
      properties:
        device_id:
          type: string
          description: Unique device identifier assigned by LLSS
          example: "dev_a1b2c3d4e5f6"
        device_secret:
          type: string
          description: >
            Secret key for device authentication. Store securely in non-volatile memory!
            This cannot be retrieved later.
          example: "xK9mN2pQ4rS6tU8vW0xY2zA4bC6dE8fG"
        auth_status:
          type: string
          enum: [pending, authorized, rejected, revoked]
          description: Current authorization status
          example: "pending"
        message:
          type: string
          description: Human-readable status message
          example: "Device registered. Waiting for admin authorization."
...
    DeviceRegistration:
      type: object
      description: Legacy device registration request (deprecated)
      required: [hardware_id, firmware_version, display]
      properties:
        hardware_id:
          type: string
        firmware_version:
          type: string
        display:
          $ref: "#/components/schemas/DisplayCapabilities"
    DeviceRegistrationResponseLegacy:
      type: object
      description: Legacy device registration response (deprecated)
      properties:
        device_id:
          type: string
        device_secret:
          type: string
        access_token:
          type: string
          description: Empty string - use auth endpoints to get tokens
    DisplayCapabilities:
      type: object
      description: Display hardware capabilities
      required: [width, height, bit_depth]
      properties:
        width:
          type: integer
          description: Display width in pixels
          example: 800
        height:
          type: integer
          description: Display height in pixels
          example: 600
        bit_depth:
          type: integer
          description: Color depth (bits per pixel, e.g., 1=B&W, 4=16 grays)
          example: 4
        partial_refresh:
          type: boolean
          default: false
          description: Whether the display supports partial refresh
    DeviceStateResponse:
      type: object
      description: Response indicating what action the device should take
      required: [action]
      properties:
        action:
          type: string
          enum: [NOOP, FETCH_FRAME, SLEEP]
          description: >
            Action for device to perform:
            - NOOP: No action needed, continue polling
            - FETCH_FRAME: New frame available, fetch it
            - SLEEP: Enter sleep mode
        frame_id:
          type: string
          nullable: true
          description: Frame ID to fetch (when action is FETCH_FRAME)
        active_instance_id:
          type: string
          nullable: true
          description: Currently active HLSS instance
        poll_after_ms:
          type: integer
          description: Suggested interval before next poll (milliseconds)
          example: 5000

    # ============================================================
    # Input Schemas
    # ============================================================
    InputEvent:
      type: object
      description: Button or input event from device
      required: [button, event_type, timestamp]
      properties:
        button:
          type: string
          enum:
            - BTN_1
            - BTN_2
            - BTN_3
            - BTN_4
            - BTN_5
            - BTN_6
            - BTN_7
            - BTN_8
            - ENTER
            - ESC
            - HL_LEFT
            - HL_RIGHT
          description: >
            Button identifier. HL_LEFT and HL_RIGHT are reserved for
            switching between assigned HLSS instances.
        event_type:
          type: string
          enum: [PRESS, LONG_PRESS, RELEASE]
          description: Type of button event
        timestamp:
          type: string
          format: date-time
          description: When the event occurred (ISO 8601)

    # ============================================================
    # Instance Schemas
    # ============================================================
    InstanceCreate:
      type: object
      description: Request to create a new HLSS instance
      required: [name, type]
      properties:
        name:
          type: string
          description: Human-readable instance name
          example: "Living Room Dashboard"
        type:
          type: string
          description: Instance type (e.g., chess, homeassistant)
          example: "homeassistant"
    Instance:
      type: object
      description: HLSS instance information
      properties:
        instance_id:
          type: string
          description: Unique instance identifier
          example: "inst_a1b2c3d4e5f6"
        name:
          type: string
          example: "Living Room Dashboard"
        type:
          type: string
          example: "homeassistant"
        access_token:
          type: string
          nullable: true
          description: Access token for HLSS to authenticate with LLSS
        created_at:
          type: string
          format: date-time

    # ============================================================
    # Frame Schemas
    # ============================================================
    FrameCreateResponse:
      type: object
      description: Response after frame submission
      properties:
        frame_id:
          type: string
          description: Unique frame identifier
          example: "frame_a1b2c3d4e5f6"
        hash:
          type: string
          description: SHA-256 hash prefix for deduplication
          example: "a1b2c3d4e5f6g7h8"
        created_at:
          type: string
          format: date-time

    # ============================================================
    # Admin Schemas
    # ============================================================
...